Tie remapping functions to event handlers

+++ lib/DacHandlerLib/DacValues.h
@@ -8,7 +8,7 @@
 #endif
 
 struct DacValues {
-    uint8_t values[4] = {0, 0, 0, 0};
+    uint16_t values[4] = {0, 0, 0, 0};
 };
 
 #endif // DacValues_h
